Refactor to separate client and proxy server in preparation for more client behaviors
